Change for 64-bit registers
Files Changed:

        * server/GdbServerImpl.cpp: (rspSyscallRequest, rspReadReg,
        rspWriteReg) Change type of register values to uint64_t.
        * targets/ITarget.h: (readRegister, writeRegister) Change
        type of value to uint64_t.
        * targets/ri5cy/Ri5cy.cpp: Likewise.
        * targets/ri5cy/Ri5cy.h: Likewise.
        * targets/ri5cy/Ri5cyImpl.cpp: Likewise, and return 8
        bytes read/written instead of 4. Also step 8 bytes in
        debug unit address space per register.
        * targets/ri5cy/Ri5cyImpl.h: Likewise.

+++ targets/ITarget.h
@@ -104,12 +104,12 @@ class ITarget
   // Read contents of a target register.
 
   virtual std::size_t  readRegister (const int  reg,
-				     uint32_t & value) const = 0;
+				     uint64_t & value) const = 0;
 
   // Write data to a target register.
 
   virtual std::size_t  writeRegister (const int  reg,
-				      const uint32_t  value) = 0;
+				      const uint64_t  value) = 0;
 
   // Read data from memory.
